requirements.txt
astroid==2.4.1
bcrypt==3.1.7
cffi==1.14.0
click==7.1.2
colorama==0.4.3
dnspython==1.16.0
email-validator==1.1.1
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.1.2
pycparser==2.20
pylint==2.5.2
PyMySQL==0.9.3
six==1.14.0
SQLAlchemy==1.3.17
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
WTForms==2.3.1
